Fixing the test.
