--- pyelftool_parser/.gitignore
@@ -1,3 +1,0 @@
-
-src/__pycache__/
-__pycache__/
