--- pyelftool_parser/testbench/global.pong/pong.ci
@@ -1,28 +1,0 @@
-graph: { title: "pong.c"
-node: { title: "pong.c:printc" label: "printc\n/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:26:1" }
-node: { title: "vsnprintf" label: "vsnprintf\n/home/minghwu/work/test/composite/src/components/lib/libc/musl-1.2.0/include/stdio.h:115:5" shape : ellipse }
-edge: { sourcename: "pong.c:printc" targetname: "vsnprintf" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:33:8" }
-node: { title: "cos_print_str" label: "cos_print_str\n/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:17:5" shape : ellipse }
-edge: { sourcename: "pong.c:printc" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:35:8" }
-node: { title: "cos_init" label: "cos_init\npong.c:23:1" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_init" targetname: "pong.c:printc" label: "pong.c:26:2" }
-node: { title: "cos_parallel_init" label: "cos_parallel_init\npong.c:31:1" }
-edge: { sourcename: "cos_parallel_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_parallel_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_parallel_init" targetname: "pong.c:printc" label: "pong.c:36:3" }
-node: { title: "parallel_main" label: "parallel_main\npong.c:44:1" }
-edge: { sourcename: "parallel_main" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "parallel_main" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "parallel_main" targetname: "pong.c:printc" label: "pong.c:48:32" }
-node: { title: "pong_call" label: "pong_call\npong.c:54:1" }
-edge: { sourcename: "pong_call" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_ret" label: "pong_ret\npong.c:62:1" }
-node: { title: "pong_arg" label: "pong_arg\npong.c:68:1" }
-node: { title: "pong_args" label: "pong_args\npong.c:74:1" }
-node: { title: "pong_wideargs" label: "pong_wideargs\npong.c:80:1" }
-node: { title: "pong_argsrets" label: "pong_argsrets\npong.c:88:1" }
-node: { title: "pong_widerets" label: "pong_widerets\npong.c:97:1" }
-node: { title: "pong_subset" label: "pong_subset\npong.c:103:1" }
-node: { title: "pong_ids" label: "pong_ids\npong.c:110:1" }
-}
