--- pyelftool_parser/testbench/ping/ping.ci
@@ -1,46 +1,0 @@
-graph: { title: "ping.c"
-node: { title: "ping.c:printc" label: "printc\n/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:26:1" }
-node: { title: "vsnprintf" label: "vsnprintf\n/home/minghwu/work/test/composite/src/components/lib/libc/musl-1.2.0/include/stdio.h:115:5" shape : ellipse }
-edge: { sourcename: "ping.c:printc" targetname: "vsnprintf" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:33:8" }
-node: { title: "cos_print_str" label: "cos_print_str\n/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:17:5" shape : ellipse }
-edge: { sourcename: "ping.c:printc" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:35:8" }
-node: { title: "cos_init" label: "cos_init\nping.c:11:1" }
-edge: { sourcename: "cos_init" targetname: "ping.c:printc" label: "ping.c:24:2" }
-node: { title: "pong_call" label: "pong_call\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:6:6" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_call" label: "ping.c:26:2" }
-node: { title: "pong_ret" label: "pong_ret\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:7:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_ret" label: "ping.c:27:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_arg" label: "pong_arg\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:8:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_arg" label: "ping.c:29:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_args" label: "pong_args\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:9:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_args" label: "ping.c:31:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_argsrets" label: "pong_argsrets\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:11:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_argsrets" label: "ping.c:33:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_wideargs" label: "pong_wideargs\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:10:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_wideargs" label: "ping.c:36:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_widerets" label: "pong_widerets\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:12:11" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_widerets" label: "ping.c:38:11" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_init" targetname: "pong_wideargs" label: "ping.c:42:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_init" targetname: "pong_widerets" label: "ping.c:44:11" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_subset" label: "pong_subset\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:13:5" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_subset" label: "ping.c:46:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-node: { title: "pong_ids" label: "pong_ids\n/home/minghwu/work/test/composite/src/components/interface/pong/./pong.h:14:9" shape : ellipse }
-edge: { sourcename: "cos_init" targetname: "pong_ids" label: "ping.c:48:8" }
-edge: { sourcename: "cos_init" targetname: "cos_print_str" label: "/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:22:9" }
-edge: { sourcename: "cos_init" targetname: "pong_call" label: "ping.c:53:3" }
-edge: { sourcename: "cos_init" targetname: "pong_argsrets" label: "ping.c:60:3" }
-node: { title: "main" label: "main\nping.c:69:1" }
-edge: { sourcename: "main" targetname: "ping.c:printc" label: "ping.c:71:2" }
-edge: { sourcename: "main" targetname: "ping.c:printc" label: "ping.c:72:2" }
-edge: { sourcename: "main" targetname: "ping.c:printc" label: "ping.c:73:2" }
-}
