+++ tools/pyelftool_parser/testbench/global.booter/llbooter.su
@@ -1,0 +1,25 @@
+/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:26:1:printc	416	static
+/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:20:1:prints.isra	32	static
+llbooter.c:71:1:boot_comp_get	16	static
+llbooter.c:103:1:execution_init	400	static
+llbooter.c:153:1:comps_init	1440	dynamic,bounded
+llbooter.c:57:1:ss_sinv_intern_is_allocated	16	static
+llbooter.c:57:1:ss_sinv_is_allocated	16	static
+llbooter.c:58:1:ss_thd_intern_is_allocated	16	static
+llbooter.c:58:1:ss_thd_is_allocated	16	static
+llbooter.c:59:1:ss_rcv_intern_is_allocated	16	static
+llbooter.c:59:1:ss_rcv_is_allocated	16	static
+llbooter.c:60:1:ss_chkpt_intern_is_allocated	16	static
+llbooter.c:60:1:ss_chkpt_is_allocated	16	static
+llbooter.c:61:1:ss_ns_asid_intern_is_allocated	16	static
+llbooter.c:61:1:ss_ns_asid_is_allocated	16	static
+llbooter.c:62:1:ss_ns_vas_intern_is_allocated	16	static
+llbooter.c:62:1:ss_ns_vas_is_allocated	16	static
+llbooter.c:452:1:chkpt_comp_init	8	static
+llbooter.c:512:1:addr_get	32	static
+llbooter.c:551:1:init_done_chkpt	576	static
+llbooter.c:596:1:init_done	16	static
+llbooter.c:615:1:init_exit	32	static
+llbooter.c:632:1:cos_parallel_init	32	static
+llbooter.c:640:1:cos_init	16	static
+llbooter.c:652:1:parallel_main	8	static
