+++ tools/pyelftool_parser/testbench/global.ping/ping.su
@@ -1,0 +1,3 @@
+/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:26:1:printc	416	static
+ping.c:11:1:cos_init	80	static
+ping.c:69:1:main	16	static
