+++ tools/pyelftool_parser/testbench/global.pong/pong.su
@@ -1,0 +1,13 @@
+/home/minghwu/work/test/composite/src/components/lib/component/./llprint.h:26:1:printc	416	static
+pong.c:23:1:cos_init	16	static
+pong.c:31:1:cos_parallel_init	32	static
+pong.c:44:1:parallel_main	16	static
+pong.c:54:1:pong_call	16	static
+pong.c:62:1:pong_ret	8	static
+pong.c:68:1:pong_arg	8	static
+pong.c:74:1:pong_args	8	static
+pong.c:80:1:pong_wideargs	8	static
+pong.c:88:1:pong_argsrets	8	static
+pong.c:97:1:pong_widerets	8	static
+pong.c:103:1:pong_subset	8	static
+pong.c:110:1:pong_ids	8	static
